Add files via upload
The project focuses on Huntington’s disease, a genetic disorder caused by a mutation in chromosome 4, specifically a repeated "CAG" nucleotide sequence encoding the amino acid glutamine. The severity of the disease is linked to the number of CAG repeats present in the patient's genes. Huntington's primarily affects individuals of European descent and has symptoms including emotional disturbances, intellectual decline, and uncontrolled movements.

The current treatments mainly address symptoms rather than curing the disease, which motivates the need for novel solutions. The project explores the use of gene editing, specifically CRISPR-Cas9, as a potential treatment. The aim is to either correct the mutated gene or prevent it from being passed down during reproduction. The experimental approach includes two strategies: non-homologous end joining and homology-directed repair, both of which involve manipulating DNA to repair mutations. However, challenges such as off-target effects, where the CRISPR system inadvertently modifies the wrong gene, and the potential for additional mutations are areas of concern that need further research.